Walmart_Backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
mysql-connector-python==8.2.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
geopy==2.4.0
python-decouple==3.8
email-validator==2.1.0
requests==2.31.0
annotated-types==0.7.0
anyio==3.7.1
bcrypt==4.3.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==45.0.5
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.1.0
exceptiongroup==1.3.0
fastapi==0.104.1
geographiclib==2.0
geopy==2.4.0
greenlet==3.2.3
h11==0.16.0
httptools==0.6.4
idna==3.10
mysql-connector-python==8.2.0
passlib==1.7.4
protobuf==4.21.12
pyasn1==0.6.1
pycparser==2.22
pydantic==2.5.0
pydantic_core==2.14.1
python-decouple==3.8
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.2
requests==2.31.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.24.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
